+++ src/Infrastructure/DoctrineMigrations/Version20200702081057.php
@@ -36,7 +36,7 @@ final class Version20200702081057 extends AbstractMigration
 
         if (!($isOracle)) {
             $this->addSql('CREATE TABLE becklyn_files (id INT AUTO_INCREMENT NOT NULL, uuid VARCHAR(36) NOT NULL, filename VARCHAR(255) NOT NULL, content_hash VARCHAR(255) NOT NULL, size INT UNSIGNED NOT NULL, owner_id VARCHAR(36) DEFAULT NULL, owner_type VARCHAR(255) DEFAULT NULL, created_ts DATETIME NOT NULL COMMENT \'(DC2Type:datetime_immutable)\', updated_ts DATETIME NOT NULL COMMENT \'(DC2Type:datetime_immutable)\', UNIQUE INDEX UNIQ_6F8BCFACD17F50A6 (uuid), PRIMARY KEY(id)) DEFAULT CHARACTER SET utf8mb4 COLLATE `utf8mb4_unicode_ci` ENGINE = InnoDB');
-            $this->addSql('CREATE TABLE becklyn_filesystem_file_pointers (id INT AUTO_INCREMENT NOT NULL, uuid VARCHAR(36) NOT NULL, file_id VARCHAR(36) NOT NULL, path VARCHAR(255) NOT NULL, created_ts DATETIME NOT NULL COMMENT \'(DC2Type:datetime_immutable)\', updated_ts DATETIME NOT NULL COMMENT \'(DC2Type:datetime_immutable)\', UNIQUE INDEX UNIQ_BF1DDD44D17F50A6 (uuid), UNIQUE INDEX UNIQ_BF1DDD4493CB796C (file_id), UNIQUE INDEX UNIQ_BF1DDD44B548B0F (path), PRIMARY KEY(id)) DEFAULT CHARACTER SET utf8mb4 COLLATE `utf8mb4_unicode_ci` ENGINE = InnoDB');
+            $this->addSql('CREATE TABLE becklyn_fs_file_pointers (id INT AUTO_INCREMENT NOT NULL, uuid VARCHAR(36) NOT NULL, file_id VARCHAR(36) NOT NULL, path VARCHAR(255) NOT NULL, created_ts DATETIME NOT NULL COMMENT \'(DC2Type:datetime_immutable)\', updated_ts DATETIME NOT NULL COMMENT \'(DC2Type:datetime_immutable)\', UNIQUE INDEX UNIQ_BF1DDD44D17F50A6 (uuid), UNIQUE INDEX UNIQ_BF1DDD4493CB796C (file_id), UNIQUE INDEX UNIQ_BF1DDD44B548B0F (path), PRIMARY KEY(id)) DEFAULT CHARACTER SET utf8mb4 COLLATE `utf8mb4_unicode_ci` ENGINE = InnoDB');
         }
 
         // Oracle gets handled in a future migration due to historical reasons
@@ -56,7 +56,7 @@ final class Version20200702081057 extends AbstractMigration
 
         if (!$isOracle) {
             $this->addSql('DROP TABLE becklyn_files');
-            $this->addSql('DROP TABLE becklyn_filesystem_file_pointers');
+            $this->addSql('DROP TABLE becklyn_fs_file_pointers');
         }
         // Oracle gets handled in a future migration due to historical reasons
     }