+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 3.0.3
 =======
 
-* (bug) Rename the File System Pointers Table to a shorter Version
+* (bc) Rename the File System Pointers Table to a shorter Version
 
 3.0.2
 =======